scripts: improve build script robustness for Xiaomi 8450 kernel

+++ build.sh
@@ -22,17 +22,17 @@ DTBO_WILDCARD="*"
 while [ "${#}" -gt 0 ]; do
     case "${1}" in
         -c | --clean )
-                DO_CLEAN=true
-                ;;
+            DO_CLEAN=true
+            ;;
         -n | --no-lto )
-                NO_LTO=true
-                ;;
+            NO_LTO=true
+            ;;
         -o | --only-config )
-                ONLY_CONFIG=true
-                ;;
+            ONLY_CONFIG=true
+            ;;
         * )
-                TARGET="${1}"
-                ;;
+            TARGET="${1}"
+            ;;
     esac
     shift
 done
@@ -48,7 +48,6 @@ if ! source .build.rc || [ -z "$SRC_ROOT" ]; then
 fi
 
 KERNEL_DIR="$SRC_ROOT/device/xiaomi/$TARGET-kernel"
-
 if [ ! -d "$KERNEL_DIR" ]; then
     echo "$KERNEL_DIR does not exist!"
     exit 1
@@ -60,12 +59,6 @@ DTBO_COPY_TO="$DTB_COPY_TO/dtbo.img"
 VBOOT_DIR="$KERNEL_DIR/vendor_ramdisk"
 VDLKM_DIR="$KERNEL_DIR/vendor_dlkm"
 
-# AK3_DIR="$HOME/AnyKernel3"
-# ZIPNAME="aospa-kernel-$TARGET-$(date '+%Y%m%d-%H%M').zip"
-# if test -z "$(git rev-parse --show-cdup 2>/dev/null)" &&
-#    head=$(git rev-parse --verify HEAD 2>/dev/null); then
-#     ZIPNAME="${ZIPNAME::-4}-$(echo $head | cut -c1-8).zip"
-# fi
 
 DEFCONFIG="gki_defconfig"
 DEFCONFIGS="vendor/waipio_GKI.config \
@@ -103,7 +96,9 @@ case "$TARGET" in
         ;;
 esac
 
+
 export PATH="$TC_DIR/bin:$PREBUILTS_DIR/bin:$PATH"
+
 
 function m() {
     make -j$(nproc --all) O=out ARCH=arm64 LLVM=1 LLVM_IAS=1 \
@@ -112,25 +107,25 @@ function m() {
         TARGET_PRODUCT=$TARGET $@ 2> >(tee -a "$LOG_FILE") || exit $?
 }
 
-$DO_CLEAN && (
+if $DO_CLEAN; then
     rm -rf out sm8450-modules
     echo "Cleaned output directories."
-)
+fi
 
 echo -e "Generating config...\n"
 mkdir -p out
 m $DEFCONFIG
 m ./scripts/kconfig/merge_config.sh $DEFCONFIGS vendor/${TARGET}_GKI.config
-scripts/config --file out/.config \
-    --set-str LOCALVERSION "-Zetta-kernel.v1"
-$NO_LTO && (
-    scripts/config --file out/.config \
-        -d LTO_CLANG_FULL -e LTO_NONE \
-    --set-str LOCALVERSION "-Zetta-kernel.v1"
-    echo -e "\nDisabled LTO!"
-)
+scripts/config --file out/.config --set-str LOCALVERSION "-Zetta-kernel.v1"
 
-$ONLY_CONFIG && exit
+if $NO_LTO; then
+    scripts/config --file out/.config -d LTO_CLANG_FULL -e LTO_NONE --set-str LOCALVERSION "-Zetta-kernel.v1-No"
+    echo -e "\nDisabled LTO!"
+fi
+
+if $ONLY_CONFIG; then
+    exit 0
+fi
 
 echo -e "\nBuilding kernel...\n"
 m Image modules dtbs
@@ -148,29 +143,15 @@ done
 echo -e "\nKernel compiled succesfully!\nMerging dtb's...\n"
 
 rm -rf out/dtbs{,-base}
-mkdir out/dtbs{,-base}
-mv  out/arch/arm64/boot/dts/vendor/qcom/$DTB_WILDCARD.dtb \
-    out/arch/arm64/boot/dts/vendor/qcom/$DTBO_WILDCARD.dtbo \
-    out/dtbs-base
+mkdir -p out/dtbs out/dtbs-base
+mv out/arch/arm64/boot/dts/vendor/qcom/$DTB_WILDCARD.dtb out/arch/arm64/boot/dts/vendor/qcom/$DTBO_WILDCARD.dtbo out/dtbs-base
 rm -f out/arch/arm64/boot/dts/vendor/qcom/*.dtbo
 ../../build/android/merge_dtbs.py out/dtbs-base out/arch/arm64/boot/dts/vendor/qcom/ out/dtbs 2> >(tee -a "$LOG_FILE") || exit $?
 
+mkdir -p "$KERNEL_COPY_TO"
+mkdir -p "$DTB_COPY_TO"
+
 echo -e "\nCopying files...\n"
-
-# rm -rf AnyKernel3
-# if [ -d "$AK3_DIR" ]; then
-#   cp -r $AK3_DIR AnyKernel3
-#   git -C AnyKernel3 checkout marble &> /dev/null
-# elif ! git clone -q https://github.com/ghostrider-reborn/AnyKernel3 -b marble; then
-#   echo -e "\nAnyKernel3 repo not found locally and couldn't clone from GitHub! Aborting..."
-#   exit 1
-# fi
-# KERNEL_COPY_TO="AnyKernel3"
-# DTB_COPY_TO="AnyKernel3/dtb"
-# DTBO_COPY_TO="AnyKernel3/dtbo.img"
-# VBOOT_DIR="AnyKernel3/vendor_boot_modules"
-# VDLKM_DIR="AnyKernel3/vendor_dlkm_modules"
-
 cp out/arch/arm64/boot/Image $KERNEL_COPY_TO
 echo "Copied kernel to $KERNEL_COPY_TO."
 
@@ -183,8 +164,9 @@ else
 fi
 echo "Copied dtb(s) to $DTB_COPY_TO."
 
-mkdtboimg.py create $DTBO_COPY_TO --page_size=4096 out/dtbs/*.dtbo 2> >(tee -a "$LOG_FILE") || exit $?
-echo "Generated dtbo.img to $DTBO_COPY_TO".
+mkdtboimg.py create --output $DTBO_COPY_TO --page_size=4096 out/dtbs/*.dtbo 2> >(tee -a "$LOG_FILE") || exit $?
+echo "Generated dtbo.img to $DTBO_COPY_TO"
+
 
 first_stage_modules="$(cat modules.list.msm.waipio)"
 second_stage_modules="$(cat modules.list.second_stage modules.list.second_stage.$TARGET)"
@@ -238,10 +220,4 @@ done
 sed -E -i 's|([^: ]*/)([^/]*\.ko)([:]?)([ ]\|$)|/lib/modules/\2\3\4|g' $VBOOT_DIR/modules.dep
 sed -E -i 's|([^: ]*/)([^/]*\.ko)([:]?)([ ]\|$)|/vendor_dlkm/lib/modules/\2\3\4|g' $VDLKM_DIR/modules.dep
 
-# cd AnyKernel3
-# zip -r9 "../$ZIPNAME" * -x .git README.md *placeholder
-# cd ..
-# rm -rf AnyKernel3
-
-echo -e "\nCompleted in $((SECONDS / 60)) minute(s) and $((SECONDS % 60)) second(s) !"
-# echo "$(realpath $ZIPNAME)"
+echo -e "\nCompleted in $((SECONDS / 60)) minute(s) and $((SECONDS % 60)) second(s)!"
